Split monolithic execution/headless/__init__.py into submodules
Extract two new submodules from the 1011-line headless/__init__.py:

- _headless_helpers.py (~175 lines): session helpers including
  _session_log_dir, _resolve_pty_mode, _resolve_session_log_dir,
  _resolve_model, _derive_step_name_from_skill_command, _recursive_snapshot,
  PostSessionMetrics, and _compute_post_session_metrics.

- _headless_execute.py (~550 lines): the _execute_claude_headless async
  function — the shared subprocess execution core for both skill sessions
  (run_headless_core) and fleet dispatch (dispatch_food_truck).

The __init__.py is now a thin facade (~419 lines) that re-exports all
public symbols from submodules while retaining run_headless_core and
DefaultHeadlessExecutor as the public API.

Backward compatibility: all existing imports from
autoskillit.execution.headless continue to work via re-exports.
