+++ src/autoskillit/execution/headless/CLAUDE.md
@@ -6,7 +6,9 @@ Headless Claude session orchestration — command prep, subprocess invocation, r
 
 | File | Purpose |
 |------|---------|
-| `__init__.py` | Main module: `run_headless_core()`, `DefaultHeadlessExecutor`, `_execute_claude_headless()` |
+| `__init__.py` | Facade: `run_headless_core()`, `DefaultHeadlessExecutor`, re-exports |
+| `_headless_helpers.py` | Session helpers: `_session_log_dir`, `_resolve_model`, `PostSessionMetrics` |
+| `_headless_execute.py` | Subprocess core: `_execute_claude_headless()` — shared skill/fleet path |
 | `_headless_git.py` | Git LOC tracking: `_capture_git_head_sha()`, `_compute_loc_changed()` |
 | `_headless_path_tokens.py` | Path-token extraction and output-path validation from assistant messages |
 | `_headless_recovery.py` | Session recovery: `_recover_from_separate_marker`, `_synthesize_from_write_artifacts` |
@@ -16,4 +18,4 @@ Headless Claude session orchestration — command prep, subprocess invocation, r
 
 ## Architecture Notes
 
-The `__init__.py` IS the main module body (not a thin facade). It uses a deferred import for `flush_session_log` to avoid circular imports. `_execute_claude_headless` is the shared path for both `run_skill` (skill session) and `dispatch_food_truck` (fleet) flows.
+The `__init__.py` is a facade with public API (`run_headless_core`, `DefaultHeadlessExecutor`) and re-exports from all submodules. `_execute_claude_headless` in `_headless_execute.py` is the shared subprocess execution path for both `run_skill` (leaf) and `dispatch_food_truck` (fleet) flows. It uses a deferred import for `flush_session_log` to avoid circular imports.
