+++ src/autoskillit/execution/headless/__init__.py
@@ -10,58 +10,47 @@ Public API:
 
 from __future__ import annotations
 
-import dataclasses
-import os
-import time
-import traceback
 from collections.abc import Callable, Mapping, Sequence
-from datetime import UTC, datetime, timedelta
 from pathlib import Path
-from typing import TYPE_CHECKING, cast
+from typing import TYPE_CHECKING
 
-import anyio
 import structlog
 
 from autoskillit.core import (
     AGENT_BACKEND_CLAUDE_CODE,
-    CAMPAIGN_ID_ENV_VAR,
-    DISPATCH_ID_ENV_VAR,
     FOOD_TRUCK_TOOL_TAGS_ENV_VAR,
-    CmdSpec,
-    CodingAgentBackend,
-    KillReason,
-    ProviderOutcome,
-    RecipeIdentity,
-    RetryReason,
     SessionCheckpoint,  # noqa: F401, TC001
     SkillResult,
     SkillSessionConfig,
     ValidatedAddDir,
     WriteBehaviorSpec,
-    claude_code_project_dir,
-    collect_version_snapshot,
     get_logger,
-    is_feature_enabled,
-    is_git_worktree,
     temp_dir_display_str,
-)
-from autoskillit.core import resolve_skill_temp_dir as _resolve_skill_temp_dir
-from autoskillit.execution.clone_guard import (
-    check_and_revert_clone_contamination,
-    is_worktree_skill,
-    snapshot_clone_state,
 )
 from autoskillit.execution.headless._headless_evidence import (
     _adapt_agent_result,  # noqa: F401
     _apply_budget_guard,  # noqa: F401
-    _build_error_path_telemetry,
-    _build_session_telemetry,
+    _build_error_path_telemetry,  # noqa: F401
+    _build_session_telemetry,  # noqa: F401
     _capture_failure,  # noqa: F401
 )
+from autoskillit.execution.headless._headless_execute import (
+    _execute_claude_headless,
+)
 from autoskillit.execution.headless._headless_git import (
-    _capture_git_head_sha,
-    _compute_loc_changed,
-    _detect_branch_divergence,
+    _capture_git_head_sha,  # noqa: F401
+    _compute_loc_changed,  # noqa: F401
+    _detect_branch_divergence,  # noqa: F401
+)
+from autoskillit.execution.headless._headless_helpers import (
+    PostSessionMetrics,
+    _compute_post_session_metrics,  # noqa: F401
+    _derive_step_name_from_skill_command,
+    _recursive_snapshot,  # noqa: F401
+    _resolve_model,
+    _resolve_pty_mode,  # noqa: F401
+    _resolve_session_log_dir,
+    _session_log_dir,  # noqa: F401
 )
 from autoskillit.execution.headless._headless_path_tokens import (  # noqa: F401
     _INTENTIONALLY_EXCLUDED_PATH_TOKENS,
@@ -78,7 +67,7 @@ from autoskillit.execution.headless._headless_recovery import (
     _CHANNEL_B_RECOVERABLE_SUBTYPES,  # noqa: F401
     _NUDGE_TIMEOUT,  # noqa: F401
     _TOKEN_NAME_RE,  # noqa: F401
-    _attempt_contract_nudge,
+    _attempt_contract_nudge,  # noqa: F401
     _extract_missing_token_hints,  # noqa: F401
     _is_path_capture_pattern,  # noqa: F401
     _merge_token_usage,  # noqa: F401
@@ -87,7 +76,7 @@ from autoskillit.execution.headless._headless_recovery import (
     _synthesize_from_write_artifacts,  # noqa: F401
 )
 from autoskillit.execution.headless._headless_result import (
-    _build_skill_result,
+    _build_skill_result,  # noqa: F401
     _parse_stdout,  # noqa: F401
     _resolve_skill_session_id,  # noqa: F401
 )
@@ -95,13 +84,7 @@ from autoskillit.execution.headless._headless_scan import _scan_jsonl_write_path
 from autoskillit.execution.recording import RecordingSubprocessRunner
 
 if TYPE_CHECKING:
-    from autoskillit.config import (
-        AutomationConfig,
-    )
-    from autoskillit.core import SubprocessResult
-    from autoskillit.pipeline.context import (
-        ToolContext,
-    )
+    from autoskillit.pipeline.context import ToolContext
 
 __all__ = [
     "DefaultHeadlessExecutor",
@@ -110,580 +93,6 @@ __all__ = [
 ]
 
 logger = get_logger(__name__)
-
-
-def _session_log_dir(cwd: str) -> Path:
-    """Derive Claude Code session log directory from project cwd.
-
-    Pre-creates the directory if absent so Channel B always has a directory
-    to poll.  Without this, a fresh clone path whose encoded project dir
-    doesn't exist yet causes ``_session_log_monitor`` to burn its entire
-    phase-1 timeout absorbing ``OSError``, ultimately producing a false
-    ``EMPTY_OUTPUT`` retry.
-    """
-    log_dir = claude_code_project_dir(cwd)
-    logger.info("session_log_dir_computed", path=str(log_dir), cwd=cwd)
-    if not log_dir.exists():
-        logger.info("session_log_dir_precreating", path=str(log_dir), cwd=cwd)
-        try:
-            log_dir.mkdir(parents=True, exist_ok=True)
-        except OSError:
-            logger.warning("session_log_dir_mkdir_failed", path=str(log_dir), cwd=cwd)
-            raise
-    return log_dir
-
-
-def _resolve_pty_mode(ctx: ToolContext) -> bool:
-    backend = cast(CodingAgentBackend, ctx.backend)
-    return backend.capabilities.pty_required
-
-
-def _resolve_session_log_dir(cwd: str, ctx: ToolContext) -> Path | None:
-    backend = cast(CodingAgentBackend, ctx.backend)
-    if not backend.capabilities.channel_b_capable:
-        return None
-    return _session_log_dir(cwd)
-
-
-def _resolve_model(
-    step_model: str,
-    config: AutomationConfig,
-    *,
-    step_name: str = "",
-    recipe_name: str = "",
-) -> str | None:
-    """Resolve model selection with 5-tier precedence.
-
-    override > recipe_overrides[recipe][step] > step_overrides[step] > step_model > default
-    """
-    if config.model.model_override:
-        logger.debug("model_resolved", tier="override", model=config.model.model_override)
-        return config.model.model_override
-    if recipe_name and step_name:
-        recipe_model = config.model.recipe_overrides.get(recipe_name, {}).get(step_name)
-        if recipe_model:
-            logger.debug("model_resolved", tier="recipe_override", model=recipe_model)
-            return recipe_model
-    if step_name:
-        step_override = config.model.step_overrides.get(step_name)
-        if step_override:
-            logger.debug("model_resolved", tier="step_override", model=step_override)
-            return step_override
-    if step_model:
-        logger.debug("model_resolved", tier="step", model=step_model)
-        return step_model
-    if config.model.default_model:
-        logger.debug("model_resolved", tier="default", model=config.model.default_model)
-        return config.model.default_model
-    logger.debug("model_resolved", tier="none", model=None)
-    return None
-
-
-def _derive_step_name_from_skill_command(skill_command: str) -> str:
-    """Extract a recording step name from a skill command string.
-
-    Examples:
-        "/autoskillit:smoke-task arg1" -> "smoke-task"
-        "/investigate foo"             -> "investigate"
-        "/autoskillit:make-plan"       -> "make-plan"
-        ""                             -> ""
-    """
-    stripped = skill_command.strip()
-    if not stripped:
-        return ""
-    token = stripped.split()[0].lstrip("/")
-    if ":" in token:
-        token = token.rsplit(":", 1)[-1]
-    return token
-
-
-def _recursive_snapshot(directory: Path) -> set[str]:
-    """Recursively enumerate all files under directory as relative paths."""
-    return {
-        str(Path(dp).relative_to(directory) / f) for dp, _, fns in os.walk(directory) for f in fns
-    }
-
-
-@dataclasses.dataclass(frozen=True, slots=True)
-class PostSessionMetrics:
-    loc_insertions: int
-    loc_deletions: int
-    effective_cwd: str
-
-
-def _compute_post_session_metrics(
-    cwd: str,
-    pre_session_sha: str,
-    skill_result: SkillResult,
-) -> PostSessionMetrics:
-    effective_cwd = skill_result.worktree_path or cwd
-    loc_ins, loc_del = _compute_loc_changed(effective_cwd, pre_session_sha)
-    return PostSessionMetrics(
-        loc_insertions=loc_ins,
-        loc_deletions=loc_del,
-        effective_cwd=effective_cwd,
-    )
-
-
-async def _execute_claude_headless(
-    spec: CmdSpec,
-    cwd: str,
-    ctx: ToolContext,
-    *,
-    skill_command: str = "",
-    step_name: str = "",
-    kitchen_id: str = "",
-    caller_session_id: str = "",
-    order_id: str = "",
-    campaign_id: str = "",
-    dispatch_id: str = "",
-    project_dir: str = "",
-    timeout: float,
-    stale_threshold: float,
-    idle_output_timeout: float | None = None,
-    expected_output_patterns: Sequence[str] = (),
-    write_behavior: WriteBehaviorSpec | None = None,
-    completion_marker: str = "",
-    prior_completion_markers: Sequence[str] | None = None,
-    recipe_name: str = "",
-    recipe_content_hash: str = "",
-    recipe_composite_hash: str = "",
-    recipe_version: str = "",
-    on_spawn: Callable[[int, int], None] | None = None,
-    skip_clone_guard: bool = False,
-    readonly_skill: bool = False,
-    write_watch_dirs: Sequence[Path] = (),
-    provider_name: str = "",
-    provider_fallback_env: dict[str, str] | None = None,
-    provider_fallback_name: str = "",
-    provider_extras: Mapping[str, str] | None = None,
-    enable_deadline_extension: bool = False,
-    max_extension_seconds: float = 7200,
-    marker_dir: Path | None = None,
-    session_id: str | None = None,
-) -> SkillResult:
-    """Shared subprocess execution for headless Claude sessions.
-
-    Accepts an already-built CmdSpec and handles runner invocation,
-    exception handling, _build_skill_result, and session log flushing.
-    Used by both run_headless_core (leaf path) and
-    DefaultHeadlessExecutor.dispatch_food_truck (food truck path).
-    """
-    campaign_id = campaign_id or os.environ.get(CAMPAIGN_ID_ENV_VAR, "")
-    dispatch_id = dispatch_id or os.environ.get(DISPATCH_ID_ENV_VAR, "")
-
-    cfg = ctx.config.run_skill
-    if idle_output_timeout is not None:
-        _raw_idle = idle_output_timeout
-    else:
-        env_idle = os.environ.get("AUTOSKILLIT_IDLE_OUTPUT_TIMEOUT")
-        if env_idle is not None:
-            try:
-                _raw_idle = float(env_idle)
-            except ValueError:
-                logger.warning(
-                    "AUTOSKILLIT_IDLE_OUTPUT_TIMEOUT: invalid float — falling back to config",
-                    env_value=env_idle,
-                    fallback=cfg.idle_output_timeout,
-                )
-                _raw_idle = float(cfg.idle_output_timeout)
-        else:
-            _raw_idle = float(cfg.idle_output_timeout)
-    effective_idle: float | None = _raw_idle if _raw_idle > 0.0 else None
-
-    current_provider_name: str = provider_name
-    fallback_activated: bool = False
-    remaining_attempts = ctx.config.providers.provider_retry_limit if provider_fallback_env else 0
-
-    runner = ctx.runner
-    if runner is None:
-        raise RuntimeError("No subprocess runner configured")
-
-    if ctx.backend is not None and spec.cmd:
-        _binary = Path(spec.cmd[0]).stem
-        if _binary in {"claude", "codex"}:
-            _expected = "claude" if ctx.backend.name == "claude-code" else "codex"
-            if _binary != _expected:
-                raise RuntimeError(
-                    f"Backend coherence violation: ctx.backend.name={ctx.backend.name!r} "
-                    f"but subprocess binary is {_binary!r}"
-                )
-
-    linux_tracing_cfg = ctx.config.linux_tracing
-    _start_ts = datetime.now(UTC).isoformat()
-    _start_mono = time.monotonic()
-    _versions = collect_version_snapshot()
-
-    _readonly_skill = readonly_skill
-    _has_write_scope = bool(write_watch_dirs)
-    _clone_snapshot = None
-    if (
-        not skip_clone_guard
-        and not is_git_worktree(Path(cwd))
-        and (is_worktree_skill(skill_command) or _readonly_skill or _has_write_scope)
-    ):
-        _clone_snapshot = await snapshot_clone_state(cwd, runner)
-
-    _watch_dirs: list[Path] = list(write_watch_dirs) if write_watch_dirs else []
-    if not _watch_dirs:
-        _default = _resolve_skill_temp_dir(cwd, skill_command)
-        if _default:
-            _watch_dirs.append(_default)
-
-    _temp_snapshots_pre: dict[Path, set[str]] = {}
-    for _wd in _watch_dirs:
-        if _wd.is_dir():
-            try:
-                _temp_snapshots_pre[_wd] = _recursive_snapshot(_wd)
-            except OSError:
-                logger.warning("watch_dir_pre_scan_failed", watch_dir=str(_wd), exc_info=True)
-                _temp_snapshots_pre[_wd] = set()
-
-    _pre_session_sha = _capture_git_head_sha(cwd)
-    _result: SubprocessResult | None = None
-    result: SubprocessResult
-    skill_result: SkillResult
-    _stream_parser = cast(CodingAgentBackend, ctx.backend).stream_parser(
-        completion_marker=completion_marker
-    )
-    while True:
-        try:
-            _result = await runner(
-                list(spec.cmd),
-                cwd=Path(cwd),
-                timeout=timeout,
-                env=spec.env,
-                pty_mode=_resolve_pty_mode(ctx),
-                session_log_dir=_resolve_session_log_dir(cwd, ctx),
-                completion_marker=completion_marker,
-                stale_threshold=stale_threshold,
-                completion_drain_timeout=cfg.completion_drain_timeout,
-                linux_tracing_config=linux_tracing_cfg,
-                idle_output_timeout=effective_idle,
-                max_suppression_seconds=cfg.max_suppression_seconds,
-                on_pid_resolved=on_spawn,
-                enable_deadline_extension=enable_deadline_extension,
-                max_extension_seconds=max_extension_seconds,
-                marker_dir=marker_dir,
-                session_id=session_id,
-                stream_parser=_stream_parser,
-            )
-        except Exception as exc:
-            logger.error("headless_runner_crashed", exc_info=True)
-            _exc_text = traceback.format_exc()
-            _log_dir = ctx.config.linux_tracing.log_dir
-            try:
-                from autoskillit.execution import flush_session_log
-
-                flush_session_log(
-                    log_dir=_log_dir,
-                    cwd=str(cwd),
-                    kitchen_id=kitchen_id,
-                    caller_session_id=caller_session_id,
-                    order_id=order_id,
-                    campaign_id=campaign_id,
-                    dispatch_id=dispatch_id,
-                    project_dir=project_dir,
-                    build_protected_campaign_ids=ctx.build_protected_campaign_ids,
-                    session_id="",
-                    pid=0,
-                    skill_command=skill_command,
-                    success=False,
-                    subtype="crashed",
-                    exit_code=-1,
-                    start_ts=_start_ts,
-                    proc_snapshots=None,
-                    termination_reason="CRASHED",
-                    exception_text=_exc_text,
-                    versions=_versions,
-                    provider_outcome=ProviderOutcome(
-                        provider_used=current_provider_name,
-                        fallback_activated=fallback_activated,
-                    ),
-                    recipe_identity=RecipeIdentity(
-                        name=recipe_name,
-                        content_hash=recipe_content_hash,
-                        composite_hash=recipe_composite_hash,
-                        version=recipe_version,
-                    ),
-                    max_sessions=ctx.config.linux_tracing.max_sessions,
-                    telemetry=_build_error_path_telemetry(ctx.github_api_log),
-                )
-            except Exception:
-                logger.debug("flush_session_log during crash failed", exc_info=True)
-            _crashed = SkillResult.crashed(
-                exception=exc,
-                skill_command=skill_command,
-                order_id=order_id,
-            )
-            return dataclasses.replace(
-                _crashed,
-                provider=ProviderOutcome(
-                    provider_used=current_provider_name, fallback_activated=fallback_activated
-                ),
-            )
-        except BaseException:
-            logger.warning("headless_runner_cancelled", exc_info=True)
-            _exc_text = traceback.format_exc()
-            _log_dir = ctx.config.linux_tracing.log_dir
-            try:
-                from autoskillit.execution import flush_session_log
-
-                with anyio.CancelScope(shield=True):
-                    flush_session_log(
-                        log_dir=_log_dir,
-                        cwd=str(cwd),
-                        kitchen_id=kitchen_id,
-                        caller_session_id=caller_session_id,
-                        order_id=order_id,
-                        campaign_id=campaign_id,
-                        dispatch_id=dispatch_id,
-                        project_dir=project_dir,
-                        build_protected_campaign_ids=ctx.build_protected_campaign_ids,
-                        session_id="",
-                        pid=0,
-                        skill_command=skill_command,
-                        success=False,
-                        subtype="cancelled",
-                        exit_code=-1,
-                        start_ts=_start_ts,
-                        proc_snapshots=None,
-                        termination_reason="CANCELLED",
-                        exception_text=_exc_text,
-                        versions=_versions,
-                        provider_outcome=ProviderOutcome(
-                            provider_used=current_provider_name,
-                            fallback_activated=fallback_activated,
-                        ),
-                        recipe_identity=RecipeIdentity(
-                            name=recipe_name,
-                            content_hash=recipe_content_hash,
-                            composite_hash=recipe_composite_hash,
-                            version=recipe_version,
-                        ),
-                        max_sessions=ctx.config.linux_tracing.max_sessions,
-                        telemetry=_build_error_path_telemetry(ctx.github_api_log),
-                    )
-            except Exception:
-                logger.debug("flush_session_log during cancel failed", exc_info=True)
-            raise
-        _elapsed = time.monotonic() - _start_mono
-        _end_ts = (datetime.fromisoformat(_start_ts) + timedelta(seconds=_elapsed)).isoformat()
-        result = dataclasses.replace(  # type: ignore[arg-type]
-            _result, start_ts=_start_ts, end_ts=_end_ts, elapsed_seconds=_elapsed
-        )
-
-        _fs_writes_detected = False
-        for _wd in _watch_dirs:
-            if _wd.is_dir():
-                try:
-                    _post = _recursive_snapshot(_wd)
-                except OSError:
-                    logger.warning("watch_dir_post_scan_failed", watch_dir=str(_wd), exc_info=True)
-                    _post = set()
-                _pre = _temp_snapshots_pre.get(_wd, set())
-                if _post - _pre:
-                    _fs_writes_detected = True
-                    break
-
-        _git_writes_detected = False
-        if is_git_worktree(Path(cwd)):
-            _git_writes_detected = _detect_branch_divergence(cwd)
-
-        audit_count_before = len(ctx.audit.get_report())
-        _supports_fmt = cast(
-            "CodingAgentBackend", ctx.backend
-        ).capabilities.supports_claude_format_stdout
-        _backend = cast(CodingAgentBackend, ctx.backend)
-        skill_result = _build_skill_result(
-            result,
-            completion_marker=completion_marker,
-            skill_command=skill_command,
-            audit=ctx.audit,
-            expected_output_patterns=expected_output_patterns,
-            cwd=cwd,
-            write_behavior=write_behavior,
-            fs_writes_detected=_fs_writes_detected,
-            git_writes_detected=_git_writes_detected,
-            prior_completion_markers=prior_completion_markers,
-            provider_used=current_provider_name,
-            supports_claude_format_stdout=_supports_fmt,
-            backend=_backend,
-        )
-
-        if (
-            skill_result.needs_retry
-            and skill_result.session_id
-            and skill_result.retry_reason
-            in (RetryReason.CONTRACT_RECOVERY, RetryReason.EARLY_STOP)
-        ):
-            nudge_success = await _attempt_contract_nudge(
-                skill_result,
-                result,
-                expected_output_patterns,
-                completion_marker,
-                cwd,
-                runner,
-                backend=ctx.backend,
-                result_parser=ctx.backend.result_parser() if ctx.backend else None,
-                provider_extras=provider_extras,
-                retry_reason=skill_result.retry_reason,
-            )
-            if nudge_success is not None:
-                skill_result = nudge_success
-
-        _clone_reverted = False
-        if _clone_snapshot is not None:
-            _effective_readonly = _readonly_skill or _has_write_scope
-            _exclude_prefix = ".autoskillit/"
-            if write_watch_dirs:
-                try:
-                    _rel = write_watch_dirs[0].relative_to(Path(cwd))
-                    _exclude_prefix = str(_rel.parts[0]) + "/"
-                except ValueError:
-                    pass  # output_dir not under cwd — keep default ".autoskillit/" fallback
-            skill_result, _clone_reverted = await check_and_revert_clone_contamination(
-                _clone_snapshot,
-                skill_result,
-                cwd,
-                runner,
-                ctx.audit,
-                skill_command=skill_command,
-                readonly_skill=_effective_readonly,
-                exclude_prefix=_exclude_prefix,
-            )
-
-        if (
-            skill_result.retry_reason in {RetryReason.STALE, RetryReason.BUDGET_EXHAUSTED}
-            and provider_fallback_env is not None
-            and remaining_attempts > 0
-            and provider_name
-            and is_feature_enabled("providers", ctx.config.features)
-        ):
-            if not fallback_activated:
-                spec = dataclasses.replace(spec, env={**spec.env, **provider_fallback_env})
-                if provider_fallback_name:
-                    current_provider_name = provider_fallback_name
-            fallback_activated = True
-            remaining_attempts -= 1
-            continue
-        break
-
-    _metrics = _compute_post_session_metrics(cwd, _pre_session_sha, skill_result)
-
-    timing_seconds: float = result.elapsed_seconds
-
-    # Extract the audit record (if any) added by this session
-    new_audit_records = ctx.audit.get_report_as_dicts()[audit_count_before:]
-    audit_record = new_audit_records[0] if new_audit_records else None
-
-    if (
-        result.proc_snapshots is not None
-        or not skill_result.success
-        or bool(step_name)
-        or skill_result.token_usage is not None
-    ):
-        from autoskillit.execution.session_log import flush_session_log
-
-        try:
-            flush_session_log(
-                log_dir=ctx.config.linux_tracing.log_dir,
-                cwd=cwd,
-                kitchen_id=kitchen_id,
-                caller_session_id=caller_session_id,
-                order_id=order_id,
-                campaign_id=campaign_id,
-                dispatch_id=dispatch_id,
-                project_dir=project_dir,
-                build_protected_campaign_ids=ctx.build_protected_campaign_ids,
-                session_id=skill_result.session_id,
-                pid=result.pid,
-                skill_command=skill_command,
-                success=skill_result.success,
-                subtype=skill_result.subtype,
-                cli_subtype=skill_result.cli_subtype,
-                exit_code=skill_result.exit_code,
-                start_ts=result.start_ts,
-                end_ts=result.end_ts,
-                elapsed_seconds=result.elapsed_seconds,
-                termination_reason=result.termination.value,
-                kill_reason=skill_result.kill_reason.value,
-                snapshot_interval_seconds=ctx.config.linux_tracing.proc_interval,
-                proc_snapshots=result.proc_snapshots,
-                step_name=step_name,
-                telemetry=_build_session_telemetry(
-                    skill_result=skill_result,
-                    timing_seconds=timing_seconds,
-                    audit_record=audit_record,
-                    github_api_log=ctx.github_api_log,
-                    loc_insertions=_metrics.loc_insertions,
-                    loc_deletions=_metrics.loc_deletions,
-                ),
-                api_retry_count=skill_result.api_retry.count,
-                api_retry_last_error=skill_result.api_retry.last_error,
-                api_retry_last_status=skill_result.api_retry.last_status,
-                api_retry_exhausted=skill_result.api_retry.exhausted,
-                write_path_warnings=skill_result.write_path_warnings,
-                write_call_count=skill_result.evidence.write_call_count,
-                fs_writes_detected=skill_result.evidence.fs_writes_detected,
-                git_writes_detected=skill_result.evidence.git_writes_detected,
-                file_changes_count=skill_result.evidence.file_changes_count,
-                clone_contamination_reverted=_clone_reverted,
-                tracked_comm=result.tracked_comm,
-                orphaned_tool_result=result.orphaned_tool_result,
-                raw_stdout=result.stdout
-                if (
-                    not skill_result.success or skill_result.kill_reason != KillReason.NATURAL_EXIT
-                )
-                else "",
-                last_stop_reason=skill_result.last_stop_reason,
-                versions=_versions,
-                provider_outcome=ProviderOutcome(
-                    provider_used=current_provider_name,
-                    fallback_activated=fallback_activated,
-                ),
-                recipe_identity=RecipeIdentity(
-                    name=recipe_name,
-                    content_hash=recipe_content_hash,
-                    composite_hash=recipe_composite_hash,
-                    version=recipe_version,
-                ),
-                max_sessions=ctx.config.linux_tracing.max_sessions,
-            )
-        except Exception:
-            logger.debug("session_log_flush_failed", exc_info=True)
-
-    logger.debug(
-        "headless_session_exit",
-        success=skill_result.success,
-        needs_retry=skill_result.needs_retry,
-        subtype=skill_result.subtype,
-        session_id=skill_result.session_id,
-    )
-
-    from autoskillit.execution.session_log import _resolve_session_label
-
-    _token_label = _resolve_session_label(step_name, dispatch_id)
-    try:
-        ctx.token_log.record(
-            _token_label,
-            skill_result.token_usage,
-            start_ts=result.start_ts,
-            end_ts=result.end_ts,
-            elapsed_seconds=result.elapsed_seconds,
-            order_id=order_id,
-            loc_insertions=_metrics.loc_insertions,
-            loc_deletions=_metrics.loc_deletions,
-        )
-    except Exception:
-        logger.debug("token_log_record_failed", exc_info=True)
-    skill_result = dataclasses.replace(
-        skill_result,
-        provider=ProviderOutcome(
-            provider_used=current_provider_name, fallback_activated=fallback_activated
-        ),
-    )
-    return skill_result
 
 
 async def run_headless_core(
