+++ tests/arch/test_subpackage_isolation.py
@@ -910,20 +910,6 @@ _LINE_LIMIT_EXEMPTIONS: dict[str, tuple[int, str]] = {
         "REQ-CNST-010-E1: canonical type registry — wide surface required to prevent "
         "circular imports; all enums/protocols/constants consolidated here",
     ),
-    "execution/headless/__init__.py": (
-        1030,
-        "REQ-CNST-010-E2: headless session orchestration — Channel B drain-race "
-        "recovery + IDLE_STALL routing + contract nudge resume tier "
-        "+ DIR_MISSING late-bind recovery arm + RecordingSubprocessRunner "
-        "step-name auto-derivation gate + recipe identity threading "
-        "+ _execute_claude_headless extraction + dispatch_food_truck orchestration-L2 path "
-        "+ campaign_id/dispatch_id propagation kwargs "
-        "+ fs-level write detection (pre/post temp-dir snapshot + _resolve_skill_temp_dir) "
-        "+ clone guard write-scope extension (planner session write isolation) "
-        "+ api_retry field forwarding to flush_session_log "
-        "+ 5-tier _resolve_model (step_overrides + recipe_overrides tiers); "
-        "splitting would fragment the adjudication pipeline across modules",
-    ),
     "session.py": (
         1060,
         "REQ-CNST-010-E3: session adjudication pipeline — exhaustive match arms "
