+++ tests/execution/test_flush_provider_integration.py
@@ -34,19 +34,19 @@ def _patch_common(monkeypatch, tmp_path, skill_result, ctx):
         return _sub_result
 
     monkeypatch.setattr(
-        "autoskillit.execution.headless._build_skill_result",
+        "autoskillit.execution.headless._headless_execute._build_skill_result",
         lambda *a, **kw: skill_result,  # noqa: ARG005
     )
     monkeypatch.setattr(
-        "autoskillit.execution.headless._compute_post_session_metrics",
+        "autoskillit.execution.headless._headless_execute._compute_post_session_metrics",
         lambda *a, **kw: PostSessionMetrics(0, 0, str(tmp_path)),  # noqa: ARG005
     )
     monkeypatch.setattr(
-        "autoskillit.execution.headless._capture_git_head_sha",
+        "autoskillit.execution.headless._headless_execute._capture_git_head_sha",
         lambda *a: "",  # noqa: ARG005
     )
     monkeypatch.setattr(
-        "autoskillit.execution.headless.collect_version_snapshot",
+        "autoskillit.execution.headless._headless_execute.collect_version_snapshot",
         lambda: {},
     )
 
@@ -126,22 +126,24 @@ class TestProviderFieldsReachFlush:
             call_count[0] += 1
             return r
 
-        monkeypatch.setattr("autoskillit.execution.headless._build_skill_result", build_result)
         monkeypatch.setattr(
-            "autoskillit.execution.headless._compute_post_session_metrics",
+            "autoskillit.execution.headless._headless_execute._build_skill_result", build_result
+        )
+        monkeypatch.setattr(
+            "autoskillit.execution.headless._headless_execute._compute_post_session_metrics",
             lambda *a, **kw: PostSessionMetrics(0, 0, str(tmp_path)),  # noqa: ARG005
         )
         monkeypatch.setattr(
-            "autoskillit.execution.headless._capture_git_head_sha",
+            "autoskillit.execution.headless._headless_execute._capture_git_head_sha",
             lambda *a: "",  # noqa: ARG005
         )
         monkeypatch.setattr(
-            "autoskillit.execution.headless.collect_version_snapshot",
+            "autoskillit.execution.headless._headless_execute.collect_version_snapshot",
             lambda: {},
         )
         monkeypatch.setattr(minimal_ctx.config.providers, "provider_retry_limit", 2)
         monkeypatch.setattr(
-            "autoskillit.execution.headless.is_feature_enabled",
+            "autoskillit.execution.headless._headless_execute.is_feature_enabled",
             lambda name, *a, **kw: name == "providers",  # noqa: ARG005
         )
 
@@ -178,7 +180,7 @@ class TestProviderFieldsReachFlush:
         from autoskillit.execution.headless import _execute_claude_headless
 
         monkeypatch.setattr(
-            "autoskillit.execution.headless.collect_version_snapshot",
+            "autoskillit.execution.headless._headless_execute.collect_version_snapshot",
             lambda: {},
         )
 
@@ -220,7 +222,7 @@ class TestProviderFieldsReachFlush:
         from autoskillit.execution.headless import _execute_claude_headless
 
         monkeypatch.setattr(
-            "autoskillit.execution.headless.collect_version_snapshot",
+            "autoskillit.execution.headless._headless_execute.collect_version_snapshot",
             lambda: {},
         )
